# PERF: GroupBy.cumsum