```
from asv_bench.benchmarks.groupby import *
self = GroupByMethods()
self.setup('int', 'cumsum', 'transformation', 2)

%timeit self.time_dtype_as_group('int', 'cumsum', 'transformation', 2)
128 µs ± 7.81 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- master
99.7 µs ± 1.13 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
```